requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
pydantic==2.5.2
gunicorn==21.2.0
